düzenleme_1

+++ ft_strnstr.c
@@ -6,15 +6,15 @@
 /*   By: aberdal <aberdal@student.42kocaeli.co      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/20 23:33:37 by aberdal           #+#    #+#             */
-/*   Updated: 2026/01/21 01:24:28 by aberdal          ###   ########.fr       */
+/*   Updated: 2026/01/26 14:24:04 by aberdal          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include "libft.h"
 
 char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
-	int		i;
-	int		j;
+	size_t		i;
+	size_t		j;
 	char	*h;
 	char	*n;
